rewrite tests to use jest
