+++ jest.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+	coverageProvider: 'v8',
+	displayName: 'unit test',
+	preset: 'ts-jest',
+	testEnvironment: 'node',
+	testRunner: 'jest-circus/runner',
+	testMatch: ['<rootDir>/tests/**/*.test.ts']
+};
