+++ package.json
@@ -4,11 +4,13 @@
 		"build": "yarn wipedist && tsc -p .",
 		"wipedist": "rimraf \"dist/\"",
 		"start": "yarn build && concurrently \"tsc -p . -w\" \"node dist/src\"",
-		"test": "yarn build && ava"
+		"test": "jest",
+		"test:watch": "jest --watch"
 	},
 	"dependencies": {
 		"@kcp/tags": "^1.0.0",
 		"@sentry/node": "5.20.1",
+		"@types/jest": "^26.0.14",
 		"@types/lodash": "^4.14.157",
 		"bufferutil": "^4.0.1",
 		"canvas": "^2.6.1",
@@ -19,6 +21,8 @@
 		"e": "^0.1.3",
 		"emoji-regex": "^9.0.0",
 		"he": "^1.2.0",
+		"jest": "^26.4.2",
+		"jest-circus": "^26.4.2",
 		"klasa": "github:dirigeants/klasa#916/head",
 		"klasa-decorators": "^0.0.1",
 		"lodash": "^4.17.19",
@@ -35,6 +39,7 @@
 		"snoostorm": "^1.2.2",
 		"snoowrap": "github:gc/snoowrap#master",
 		"threads": "^1.4.1",
+		"ts-jest": "^26.4.0",
 		"ts-toolbelt": "^6.9.9",
 		"twit": "^2.2.9",
 		"typeorm": "^0.2.25",
